+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org).
 ## Unreleased
 - Added - cli.success() and cli.fail() overloads with just the exit code
 - Added - cli.parse() const overloads for vector and argv
+- Added - priority parameter when adding actions
 
 ## dimcli 7.4.0 (2025-06-16)
 - Added - cli.beforeExec() and cli.afterExec() for code common to all commands